+++ neoload-project/src/main/java/com/neotys/neoload/model/repository/RandomNumberVariable.java
@@ -4,6 +4,6 @@ import org.immutables.value.Value;
 
 @Value.Immutable
 public interface RandomNumberVariable extends Variable{
-	int getMinValue();
-	int getMaxValue();
+	long getMinValue();
+	long getMaxValue();
 }
